# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Document": {
        "description": "An uploaded original; identical bytes always map to one document.",
        "properties": {
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "media_type": {
            "const": "text/plain",
            "default": "text/plain",
            "title": "Media Type",
            "type": "string"
          },
          "size_bytes": {
            "minimum": 1.0,
            "title": "Size Bytes",
            "type": "integer"
          },
          "text_unit_count": {
            "minimum": 1.0,
            "title": "Text Unit Count",
            "type": "integer"
          }
        },
        "required": [
          "size_bytes",
          "content_hash",
          "text_unit_count"
        ],
        "title": "Document",
        "type": "object"
      },
...
      "ExtractionStatus": {
        "properties": {
          "extracted": {
            "title": "Extracted",
            "type": "integer"
          },
          "status": {
            "enum": [
              "not_started",
              "running",
              "failed",
              "complete"
            ],
            "title": "Status",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "extracted",
          "total"
        ],
        "title": "ExtractionStatus",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JsonValue": {},
...
      "TextUnit": {
        "properties": {
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "end_offset": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Offset"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "media_type": {
            "minLength": 1,
            "title": "Media Type",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Metadata",
            "type": "object"
          },
          "source_document_id": {
            "format": "uuid",
            "title": "Source Document Id",
            "type": "string"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "start_offset": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Offset"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "token_count": {
            "minimum": 0.0,
            "title": "Token Count",
            "type": "integer"
          }
        },
        "required": [
          "text",
          "source_document_id",
          "token_count",
          "media_type",
          "content_hash"
        ],
        "title": "TextUnit",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Aether",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v0/documents": {
      "post": {
        "description": "Upload a UTF-8 text/plain body, at most 1 MiB. Re-uploads return the original.",
        "operationId": "upload_document_api_v0_documents_post",
        "parameters": [
          {
            "in": "query",
            "name": "filename",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filename"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Same bytes already uploaded"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Database unavailable"
          }
        },
        "summary": "Upload Document"
      }
    },
    "/api/v0/documents/{document_id}": {
      "get": {
        "operationId": "get_document_api_v0_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Database unavailable"
          }
        },
        "summary": "Get Document"
      }
    },
    "/api/v0/documents/{document_id}/extraction": {
      "get": {
        "operationId": "get_extraction_api_v0_documents__document_id__extraction_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Database unavailable"
          }
        },
        "summary": "Get Extraction"
      },
      "post": {
        "description": "Queue LLM extraction. Idempotent: finished units are skipped.",
        "operationId": "start_extraction_api_v0_documents__document_id__extraction_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Database unavailable"
          }
        },
        "summary": "Start Extraction"
      }
    },
    "/api/v0/documents/{document_id}/text-units": {
      "get": {
        "operationId": "list_text_units_api_v0_documents__document_id__text_units_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TextUnit"
                  },
                  "title": "Response List Text Units Api V0 Documents  Document Id  Text Units Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Database unavailable"
          }
        },
        "summary": "List Text Units"
      }
    },
...
  }
}